chore: rename command directory for auth things
