+++ src/commands/auth/registerUser.js
